refactor: SearchPage to use hooks
