+++ src/core_modules/capture-core/components/Pages/Search/SearchDomainSelector/SearchDomainSelector.js
@@ -1,0 +1,88 @@
+// @flow
+import React, { useMemo } from 'react';
+import i18n from '@dhis2/d2-i18n';
+import withStyles from '@material-ui/core/styles/withStyles';
+import {
+    SingleSelect,
+    SingleSelectOption,
+} from '@dhis2/ui-core';
+import { Section, SectionHeaderSimple } from '../../../Section';
+import type { Props } from './SearchDomainSelector.types';
+
+const styles = (theme: Theme) => ({
+    searchDomainSelectorSection: {
+        maxWidth: theme.typography.pxToRem(900),
+        marginBottom: theme.typography.pxToRem(20),
+    },
+    searchRow: {
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    searchRowTitle: {
+        flexBasis: 200,
+        marginLeft: 8,
+    },
+    searchRowSelectElement: {
+        width: '100%',
+    },
+    customEmpty: {
+        textAlign: 'center',
+        padding: '8px 24px',
+    },
+    divider: {
+        padding: '8px',
+    },
+});
+
+export const Index =
+  ({ trackedEntityTypesWithCorrelatedPrograms, classes, onSelect, selectedProgram }: Props) =>
+      (<Section
+          className={classes.searchDomainSelectorSection}
+          header={
+              <SectionHeaderSimple
+                  containerStyle={{ paddingLeft: 8, borderBottom: '1px solid #ECEFF1' }}
+                  title={i18n.t('Search')}
+              />
+          }
+      >
+          <div className={classes.searchRow} style={{ padding: '8px 0' }}>
+              <div className={classes.searchRowTitle}>Search for</div>
+              <div className={classes.searchRowSelectElement} style={{ marginRight: 8 }}>
+                  <SingleSelect
+                      onChange={({ selected }) => { onSelect(selected); }}
+                      selected={selectedProgram}
+                      empty={<div className={classes.customEmpty}>Custom empty component</div>}
+                  >
+                      {
+                          useMemo(() => Object.values(trackedEntityTypesWithCorrelatedPrograms)
+                              // $FlowFixMe https://github.com/facebook/flow/issues/2221
+                              .map(({ trackedEntityTypeName, trackedEntityTypeId, programs: tePrograms }) =>
+                                  // SingleSelect component wont allow us to wrap the SingleSelectOption
+                                  // in any other element and still make use of the default behaviour.
+                                  // Therefore we are returning the group title and the
+                                  // SingleSelectOption in an array.
+                                  [
+                                      <SingleSelectOption
+                                          value={trackedEntityTypeId}
+                                          label={trackedEntityTypeName}
+                                      />,
+                                      tePrograms.map(({ programName, programId }) =>
+                                          (<SingleSelectOption value={programId} label={programName} />)),
+                                      <div className={classes.divider} key={trackedEntityTypeId}>
+                                          <hr />
+                                      </div>,
+                                  ],
+                              ),
+                          [
+                              classes.divider,
+                              trackedEntityTypesWithCorrelatedPrograms,
+                          ])
+                      }
+                  </SingleSelect>
+              </div>
+          </div>
+      </Section>
+      );
+
+export const SearchDomainSelector = withStyles(styles)(Index);
