+++ src/core_modules/capture-core/components/Pages/Search/SearchDomainSelector/index.js
@@ -1,0 +1,1 @@
+export { SearchDomainSelector } from './SearchDomainSelector';
