+++ src/core_modules/capture-core/components/Pages/Search/SearchPage.actions.js
@@ -25,3 +25,7 @@ export const searchViaAttributesOnScopeTrackedEntityType = ({ trackedEntityTypeI
 
 export const searchViaAttributesOnScopeProgram = ({ programId, formId, page = 1 }) =>
     actionCreator(searchPageActionTypes.VIA_ATTRIBUTES_ON_SCOPE_PROGRAM_SEARCH)({ programId, formId, page });
+
+export const showInitialSearchPage = () => actionCreator(searchPageActionTypes.SEARCH_RESULTS_INITIAL_VIEW)();
+
+export const navigateToMainPage = () => actionCreator(searchPageActionTypes.TO_MAIN_PAGE_NAVIGATE)();
