+++ src/core_modules/capture-core/components/Pages/Search/SearchPage.container.js
@@ -2,40 +2,8 @@
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { SearchPage as SearchPageComponent } from './SearchPage.component';
-import { programCollection } from '../../../metaDataMemoryStores';
-import { TrackerProgram } from '../../../metaData/Program';
 import { withErrorMessageHandler, withLoadingIndicator } from '../../../HOC';
-import type { DispatchersFromRedux, OwnProps, Props, PropsFromRedux } from './SearchPage.types';
-import { addFormData } from '../../D2Form/actions/form.actions';
-import { actionCreator } from '../../../actions/actions.utils';
-import { searchPageActionTypes } from './SearchPage.actions';
-
-const buildSearchOption = (id, name, searchGroups, searchScope) => ({
-    searchOptionId: id,
-    searchOptionName: name,
-    searchGroups: [...searchGroups.values()]
-        // We use the sorted array to always have expanded the first search group section.
-        .sort(({ unique: xBoolean }, { unique: yBoolean }) => {
-            if (xBoolean === yBoolean) {
-                return 0;
-            }
-            if (xBoolean) {
-                return -1;
-            }
-            return 1;
-        })
-        .map(({ unique, searchForm, minAttributesRequiredToSearch }, index) => ({
-            unique,
-            searchForm,
-            // We adding the `formId` here for the reason that we will use it in the SearchPage component.
-            // Specifically the function `addFormData` will add an object for each input field to the store.
-            // Also the formId is passed in the `Form` component and needs to be identical with the one in
-            // the store in order for the `Form` to function. For these reasons we generate it once here.
-            formId: `searchPageForm-${id}-${index}`,
-            searchScope,
-            minAttributesRequiredToSearch,
-        })),
-});
+import type { OwnProps, Props, PropsFromRedux } from './SearchPage.types';
 
 export const searchScopes = {
     PROGRAM: 'PROGRAM',
@@ -43,79 +11,17 @@ export const searchScopes = {
 };
 
 const mapStateToProps = (state: ReduxState): PropsFromRedux => {
-    const { currentSelections, activePage, searchPage: { searchStatus, generalPurposeErrorMessage } } = state;
-
-    const trackedEntityTypesWithCorrelatedPrograms =
-      [...programCollection.values()]
-          .filter(program => program instanceof TrackerProgram)
-          // $FlowFixMe
-          .reduce((acc, {
-              id: programId,
-              name: programName,
-              trackedEntityType: {
-                  id: trackedEntityTypeId,
-                  name: trackedEntityTypeName,
-                  searchGroups: trackedEntityTypeSearchGroups,
-              },
-              searchGroups,
-          }: TrackerProgram) => {
-              const accumulatedProgramsOfTrackedEntityType =
-                acc[trackedEntityTypeId] ? acc[trackedEntityTypeId].programs : [];
-              return {
-                  ...acc,
-                  [trackedEntityTypeId]: {
-                      trackedEntityTypeId,
-                      trackedEntityTypeName,
-                      trackedEntityTypeSearchGroups,
-                      programs: [
-                          ...accumulatedProgramsOfTrackedEntityType,
-                          { programId, programName, searchGroups },
-                      ],
-
-                  },
-              };
-          }, {});
-
-    const preselectedProgram = Object.values(trackedEntityTypesWithCorrelatedPrograms)
-        // $FlowFixMe https://github.com/facebook/flow/issues/2221
-        .map(({ programs }) =>
-            programs.find(({ programId }) => programId === currentSelections.programId))
-        .filter(program => program)[0];
-
-    const availableSearchOptions = Object.values(trackedEntityTypesWithCorrelatedPrograms)
-        // $FlowFixMe https://github.com/facebook/flow/issues/2221
-        .reduce((acc, { trackedEntityTypeId, trackedEntityTypeName, trackedEntityTypeSearchGroups, programs }) => ({
-            ...acc,
-            [trackedEntityTypeId]: buildSearchOption(trackedEntityTypeId, trackedEntityTypeName, trackedEntityTypeSearchGroups, searchScopes.TRACKED_ENTITY_TYPE),
-            ...programs.reduce((accumulated, { programId, programName, searchGroups }) => ({
-                ...accumulated,
-                [programId]: buildSearchOption(programId, programName, searchGroups, searchScopes.PROGRAM),
-            }), {}),
-        }), {});
+    const { activePage } = state;
 
     return {
-        preselectedProgram: {
-            value: preselectedProgram && preselectedProgram.programId,
-            label: preselectedProgram && preselectedProgram.programName,
-        },
-        availableSearchOptions,
-        trackedEntityTypesWithCorrelatedPrograms,
         error: activePage.selectionsError && activePage.selectionsError.error,
         ready: !activePage.isLoading,
-        searchStatus,
-        generalPurposeErrorMessage,
     };
 };
 
-const mapDispatchToProps = (dispatch: ReduxDispatch): DispatchersFromRedux => ({
-    addFormIdToReduxStore: (formId) => { dispatch(addFormData(formId)); },
-    showInitialSearchPage: () => { dispatch(actionCreator(searchPageActionTypes.SEARCH_RESULTS_INITIAL_VIEW)()); },
-    navigateToMainPage: () => { dispatch(actionCreator(searchPageActionTypes.TO_MAIN_PAGE_NAVIGATE)()); },
-});
-
 export const SearchPage =
   compose(
-      connect<Props, OwnProps, _, _, _, _>(mapStateToProps, mapDispatchToProps),
+      connect<Props, OwnProps, _, _, _, _>(mapStateToProps, () => null),
       withLoadingIndicator(),
       withErrorMessageHandler(),
   )(SearchPageComponent);
